# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "可查詢weblog的資訊",
        "title": "weblog API",
        "contact": {},
        "license": {
            "name": "Apache 2.0"
        },
        "version": "1.0"
    },
    "paths": {
        "/log/logs": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Get weblog.查詢"
                ],
                "summary": "查詢所有weblog資訊",
                "responses": {}
            },
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "新增"
                ],
                "summary": "ADD新增",
                "parameters": [
                    {
                        "description": "data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pojo.Log"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/log/logs/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Get weblog.查詢"
                ],
                "summary": "以Id搜尋welog資訊",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        }
    },
    "definitions": {
        "pojo.Log": {
            "type": "object",
            "properties": {
                "logId": {
                    "description": "定義json的name是什麼",
                    "type": "integer"
                },
                "logip": {
                    "type": "string"
                },
                "logstatus": {
                    "type": "string"
                },
                "logurl": {
                    "type": "string"
                }
            }
        }
    }
}